Bug 781731 - Fixed failed assertion in reftest1; r=doublec

+++ gfx/2d/Blur.cpp
@@ -387,7 +387,7 @@ AlphaBoxBlur::AlphaBoxBlur(const Rect& aRect,
 
 AlphaBoxBlur::AlphaBoxBlur(uint8_t* aData,
                            const Rect& aRect,
-                           int8_t aStride,
+                           int32_t aStride,
                            float aSigma)
   : mSpreadRadius(),
     mBlurRadius(CalculateBlurRadius(Point(aSigma, aSigma))),
